Rename brille_interpolator.py to brille.py
For less verbose importing
